refactor: [#1203] move test
